move UnreadSourceAccessor to its package
