+++ linen-modern/src/main/scala/x7c1/linen/modern/action/DetailAreaAction.scala
@@ -1,6 +1,7 @@
 package x7c1.linen.modern.action
 
-import x7c1.linen.modern.accessor.{RawSourceAccessor, UnreadSourceAccessor, EntryAccessor}
+import x7c1.linen.modern.accessor.unread.UnreadSourceAccessor
+import x7c1.linen.modern.accessor.{RawSourceAccessor, EntryAccessor}
 import x7c1.linen.modern.display.unread.{OutlineSelectedEvent, DetailSelectedEvent, DetailArea, SourceSelectedEvent}
 import x7c1.linen.modern.struct.UnreadDetail
 import x7c1.wheat.modern.callback.CallbackTask.task
