+++ linen-modern/src/main/scala/x7c1/linen/modern/action/OutlineAreaAction.scala
@@ -1,6 +1,7 @@
 package x7c1.linen.modern.action
 
-import x7c1.linen.modern.accessor.{RawSourceAccessor, EntryAccessor, UnreadSourceAccessor}
+import x7c1.linen.modern.accessor.unread.UnreadSourceAccessor
+import x7c1.linen.modern.accessor.{RawSourceAccessor, EntryAccessor}
 import x7c1.linen.modern.display.unread.{OutlineSelectedEvent, DetailSelectedEvent, OutlineArea, SourceSelectedEvent}
 import x7c1.linen.modern.struct.UnreadOutline
 import x7c1.wheat.modern.callback.CallbackTask
