+++ linen-modern/src/main/scala/x7c1/linen/modern/action/SourceAreaAction.scala
@@ -1,6 +1,6 @@
 package x7c1.linen.modern.action
 
-import x7c1.linen.modern.accessor.UnreadSourceAccessor
+import x7c1.linen.modern.accessor.unread.UnreadSourceAccessor
 import x7c1.linen.modern.display.unread.{OutlineSelectedEvent, DetailSelectedEvent, SourceSelectedEvent, SourceArea}
 import x7c1.wheat.modern.callback.CallbackTask.task
 import x7c1.wheat.modern.tasks.Async.await
