+++ linen-modern/src/main/scala/x7c1/linen/modern/display/unread/SourceArea.scala
@@ -1,7 +1,7 @@
 package x7c1.linen.modern.display.unread
 
 import android.support.v7.widget.RecyclerView
-import x7c1.linen.modern.accessor.UnreadSourceAccessor
+import x7c1.linen.modern.accessor.unread.UnreadSourceAccessor
 import x7c1.wheat.modern.tasks.ScrollerTasks
