+++ linen-modern/src/main/scala/x7c1/linen/modern/display/unread/SourceRowAdapter.scala
@@ -3,7 +3,7 @@ package x7c1.linen.modern.display.unread
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.ViewGroup
 import x7c1.linen.glue.res.layout.{UnreadSourceRowFooter, UnreadSourceRowItem, UnreadSourceRow}
-import x7c1.linen.modern.accessor.UnreadSourceAccessor
+import x7c1.linen.modern.accessor.unread.UnreadSourceAccessor
 import x7c1.linen.modern.struct.UnreadSource
 import x7c1.wheat.ancient.resource.ViewHolderProvider
 import x7c1.wheat.macros.logger.Log
