+++ linen-modern/src/main/scala/x7c1/linen/modern/init/unread/AccessorLoader.scala
@@ -4,7 +4,8 @@ import android.app.LoaderManager
 import android.database.sqlite.SQLiteDatabase
 import x7c1.linen.glue.res.layout.MainLayout
 import x7c1.linen.modern.accessor.ChannelAccessor.findCurrentChannelId
-import x7c1.linen.modern.accessor.{AccountAccessor, AccountIdentifiable, EntryAccessor, EntryAccessorBinder, UnreadSourceAccessor}
+import x7c1.linen.modern.accessor.unread.UnreadSourceAccessor
+import x7c1.linen.modern.accessor.{AccountAccessor, AccountIdentifiable, EntryAccessor, EntryAccessorBinder}
 import x7c1.linen.modern.init.unread.AccessorLoader.inspectSourceAccessor
 import x7c1.linen.modern.init.unread.SourceNotLoaded.{AccountNotFound, Abort, ChannelNotFound, ErrorEmpty}
 import x7c1.linen.modern.struct.{UnreadDetail, UnreadOutline, UnreadSource}
