+++ linen-modern/src/main/scala/x7c1/linen/modern/init/unread/UnreadItemsDelegatee.scala
@@ -8,7 +8,8 @@ import android.view.KeyEvent
 import x7c1.linen.glue.activity.ActivityControl
 import x7c1.linen.glue.res.layout.{UnreadDetailRowFooter, UnreadOutlineRowFooter, UnreadSourceRowFooter, MainLayout, MenuRowLabel, MenuRowSeparator, MenuRowTitle, UnreadDetailRowEntry, UnreadDetailRowSource, UnreadOutlineRowEntry, UnreadOutlineRowSource, UnreadSourceRowItem}
 import x7c1.linen.modern.accessor.preset.{ClientAccount, ClientAccountSetup, PresetAccount}
-import x7c1.linen.modern.accessor.{EntryAccessor, LinenOpenHelper, RawSourceAccessor, UnreadSourceAccessor}
+import x7c1.linen.modern.accessor.unread.UnreadSourceAccessor
+import x7c1.linen.modern.accessor.{EntryAccessor, LinenOpenHelper, RawSourceAccessor}
 import x7c1.linen.modern.display.unread.{DetailArea, OutlineArea, PaneContainer, SourceArea}
 import x7c1.linen.modern.struct.{UnreadDetail, UnreadOutline}
 import x7c1.wheat.ancient.resource.ViewHolderProvider
