+++ linen-modern/src/test/scala/x7c1/linen/modern/accessor/SourceOpenHelperTest.scala
@@ -7,7 +7,7 @@ import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
-import x7c1.linen.modern.accessor.unread.UnreadSourceAccessorQueries
+import x7c1.linen.modern.accessor.unread.{UnreadSourceAccessor, UnreadSourceAccessorQueries}
 import x7c1.linen.modern.init.dev.DummyFactory
 import x7c1.linen.modern.init.unread.AccessorLoader
